src/main/java/com/adira/contact/common/Utils.java
package com.adira.contact.common;

import java.util.List;
import java.util.UUID;
import java.util.stream.Collectors;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.BindingResult;
import org.springframework.validation.ObjectError;

import com.adira.contact.entity.ApiResponse;

public class Utils {

    public boolean validatePassword(String password) {
        return password.matches("^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$");
    }

    public ResponseEntity<ApiResponse<?>> handleValidationErrors(BindingResult bindingResult) {
        List<String> errors = bindingResult.getAllErrors()
                .stream()
                .map(ObjectError::getDefaultMessage)
                .collect(Collectors.toList());

        return ResponseEntity.badRequest()
                .body(new ApiResponse<>(HttpStatus.BAD_REQUEST.value(), "Validation failed", "API", errors));
    }

    public String usingRandomUUID() {

        UUID randomUUID = UUID.randomUUID();

        return randomUUID.toString().replaceAll("_", "");

    }

    //Fields should not be empty or contain only whitespaces, true --> Error
    public boolean isNonEmptyString(String value) {
        if (value == null){
            return false;
        }

        return value.trim().isEmpty();
        
    }

    
}
